client: make it possible to start mpv by API
This involves messing with lots of weird details in the init code.
Although it would be nicer to have the console player's main function
use the public client API, but this turned out as too complicated. It's
better to keep these code paths separate (although most code is still
shared between them).

This is not complete yet. There are some omissions too.