client: allow setting a custom wakeup callback
Not sure if this is a good idea. But it sure might help the API user to
keep the client wait loop simple. For example, the wakeup callback could
be used to wake up the GUI event loop in order to make it query events
from the client API.

The possible "bad idea" in this is that the user is given a better
chance to mess up the mpv core (by blocking in the callback), or the
client API (because the callback is called from a foreign,
unsynchronized thread).